added routes to users, map api
