+++ api/desktop/index.js
@@ -1,3 +1,5 @@
 const express = require('express');
 const router = express.Router();
 const { Comments, Images, Pins, Projects, Schematics, Users } = require('../../models');
+
+module.exports = router;
