+++ server.js
@@ -16,127 +16,6 @@ app.use(express.static('public'));
 app.use(bp.json());
 app.use('/api', require('./api/index.js'));
 
-app.get("/projects", (req,res) => {
-   Projects.findAll({
-    include: [
-      {model: Schematics,
-        include: [{ model: Pins,
-          include: [{ model: Images},{ model: Comments}]
-         }]
-      }
-    ]
-   })
-      .then(project => {
-        function result(){
-          return  project.map(proj => {
-
-            return {
-              project:{
-                  id: proj.id,
-                  title: proj.title,
-                  address: proj.address,
-                  job_number: proj.job_number,
-                  client_name: proj.client_name,
-                  updatedAt: proj.updatedAt,
-                  createdAt: proj.createdAt
-             },
-             schematic: {
-                  id: proj.Schematic.id,
-                  image_url: proj.Schematic.image_url,
-                  updatedAt: proj.Schematic.updatedAt,
-                  createdAt: proj.Schematic.createdAt
-             },
-             pin: {
-                id: proj.Schematic.Pins
-             }
-
-            };
-          });
-
-
-        }
-        res.json(result());
-      });
-});
-
-app.get("/schematics", (req,res) => {
-   Schematics.findAll({
-    include: [{model: Pins}]
-   })
-      .then(schematic => {
-        res.json(schematic);
-      });
-});
-
-app.post("/projects", (req, res) => {
-  Projects.create({
-    title: req.body.title,
-    address: req.body.address,
-    client_name: req.body.client_name,
-    job_number: req.body.job_number,
-  }).then((project) => {
-      res.json(project.dataValues);
-   })
-  .catch((err) => {
-    console.log(err);
-  });
-});
-
-
-
-app.post('/schematics', (req, res) => {
-  return Schematics.create({
-    image_url: req.body.image_url,
-    project_id: req.body.project_id
-  })
-  .then( (schematic) => {
-    return res.json(schematic);
-  });
-});
-
-app.post('/pins', (req, res) => {
-  return Pins.create({
-    x: req.body.x,
-    y: req.body.y,
-    isActive: req.body.isActive,
-    width: req.body.width,
-    height: req.body.height,
-    isPositionOutside: req.body.isPositionOutside,
-    isMouseDetected: req.body.isMouseDetected,
-    isTouchDetected: req.body.isTouchDetected,
-    schematic_id: req.body.schematic_id
-  })
-  .then( (pin) => {
-    return res.json(pin);
-  })
-  .catch( (err) => {
-    console.log(err);
-  });
-
-});
-
-app.post('/images', (req, res) => {
-  return Images.create({
-    image_url: req.body.image_url,
-    pin_id: req.body.pin_id
-  })
-  .then( (image) => {
-    return res.json(image);
-  });
-});
-
-app.post('/comments', (req, res) => {
-  return Comments.create({
-    body: req.body.body,
-    pin_id: req.body.pin_id
-
-  })
-  .then( (body) => {
-    return res.json(body);
-  });
-});
-
-
 
 app.get('*', (req, res) => {
   res.sendFile('./public/index.html', { root: __dirname });
